[Fix] 히어독일 때 시그널 표시되는 문제 수정
